+++ src/components/EntryEditor.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 import ImmutablePropTypes from 'react-immutable-proptypes';
-import ControlPane from './ControlPane';
+import ControlPane from './ControlPanel/ControlPane';
 import PreviewPane from './PreviewPane';
 import styles from './EntryEditor.css';
