+++ src/components/Widgets/MarkdownControlElements/RawEditor/index.css
@@ -1,0 +1,13 @@
+.root {
+  font-family: monospace;
+  display: block;
+  width: 100%;
+  padding: 0;
+  margin: 0;
+  border: none;
+  outline: 0;
+  box-shadow: none;
+  background: 0 0;
+  font-size: 18px;
+  color: #7c8382;
+}
